fix v1.0.0 buildability: add missing event_groups.c, cast pointer to uint32_t
v1.0.0 depended on Middlewares/stm32f746_lcd_test/FreeRTOS/Core/Src/event_groups.c,
which existed locally but was never committed -- linking fails with "undefined
reference to xEventGroupSetBits" etc. without it. Also add the (uint32_t) cast
in MY_Front_LCD_DrawBitmap's pointer arithmetic, which recent GCC (14.x) treats
as a hard error rather than a warning. Neither change alters behavior; this is
the actual buildable baseline going forward. v1.0.0 itself is left untouched
as the historical snapshot.

+++ Drivers/MY_DRIVER/stream_system.c
@@ -389,7 +389,7 @@ void MY_Front_LCD_DrawBitmap(uint32_t Xpos, uint32_t Ypos, uint8_t *pbmp, uint8_
   bit_pixel = pbmp[28] + (pbmp[29] << 8);  
   
   /* Set the address */
-  address = layer1_buff[buff_idx] + (((BSP_LCD_GetXSize()*Ypos) + Xpos)*(4));
+  address = (uint32_t)layer1_buff[buff_idx] + (((BSP_LCD_GetXSize()*Ypos) + Xpos)*(4));
   
   /* Bypass the bitmap header */
   pbmp += (index + (width * (height - 1) * (bit_pixel/8)));
